Move logger to common module.
