+++ packages/server/customer-os-api/.env
@@ -30,7 +30,7 @@ EVENTS_PROCESSING_PLATFORM_API_KEY=082c1193-a5a2-42fc-87fc-e960e692fffd
 
 # valid log values (PANIC, FATAL, ERROR, WARN, INFO, DEBUG)
 LOGGER_LEVEL: DEBUG
-LOGGER_DEV_MODE: false
+LOGGER_DEV_MODE: true
 LOGGER_ENCODER: console
 
 ADMIN_KEY: 13604691-6850-4f08-a568-d995fb57b4f3
