+++ packages/server/customer-os-api/customer-os-api.go
@@ -4,8 +4,8 @@ import (
 	"context"
 	"github.com/openline-ai/openline-customer-os/packages/server/customer-os-api/config"
 	"github.com/openline-ai/openline-customer-os/packages/server/customer-os-api/constants"
-	"github.com/openline-ai/openline-customer-os/packages/server/customer-os-api/logger"
 	"github.com/openline-ai/openline-customer-os/packages/server/customer-os-api/server"
+	"github.com/openline-ai/openline-customer-os/packages/server/customer-os-common-module/logger"
 	"log"
 	"os"
 	"os/signal"
