+++ packages/server/customer-os-api/graph/resolver/resolver.go
@@ -2,8 +2,8 @@ package resolver
 
 import (
 	"github.com/openline-ai/openline-customer-os/packages/server/customer-os-api/grpc_client"
-	"github.com/openline-ai/openline-customer-os/packages/server/customer-os-api/logger"
 	"github.com/openline-ai/openline-customer-os/packages/server/customer-os-api/service"
+	"github.com/openline-ai/openline-customer-os/packages/server/customer-os-common-module/logger"
 )
 
 //go:generate go run github.com/99designs/gqlgen
